// src/api/openapi/model/calendarEntry.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserEssential } from './userEssential';
import { Calendar } from './calendar';
import { Lock } from './lock';


export interface CalendarEntry { 
    lock: Lock;
    title: any | null;
    description: any | null;
    start_time: any | null;
    end_time: any | null;
    id: any | null;
    timestamp: any | null;
    user: UserEssential;
    calendar: Calendar;
}