// src/api/openapi/model/calendarEntryCreate.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CalendarEntryCreate { 
    title: any | null;
    description: any | null;
    start_time: any | null;
    end_time: any | null;
    company?: any | null;
}